# bookingAPI/swagger/bookingapi.yml
swagger: '2.0'
info:
  description: Walks Booking API 
  version: 1.0.0
  title: Walks Booking API
  contact:
    email: admins@walks.org
host: staging-bookingapi.walks.org
basePath: /v1
...
schemes:
  - https
paths:
...
  /booking/{booking_id}/promo:
    delete:
      tags:
        - promo
      summary: Remove promo from clients cart total 
      description: ''
      operationId: removePromo
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: booking_id
          in: path
          required: true
          type: integer
          minimum: 1
          description: The booking id
      responses:
        '200':
          $ref: '#/responses/Success'
        '401':
          $ref: '#/responses/Unauthorized'

##################################
## Misc Tag
##################################
...
responses:
  Success:
    description: Success
    schema:
      $ref: '#/definitions/Success'
...
  Unauthorized:
    description: Unauthorized
    schema:
      $ref: '#/definitions/Error'
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header


##################################
## Parameter Definitions
## Input/Output
##################################
definitions:
...
  Success:
    type: object
    properties:
      code:
        type: string
      data:
        type: string
      message:
        type: string
    required:
      - code
      - message
  Error:
    type: object
    properties:
      code:
        type: string
      message:
        type: string
    required:
      - code
      - message
...